freebsd: Add support for FreeBSD 12.x ABI
In FreeBSD 12.x, several ABI were changed in an incompatible way:

* Inodes were changed from 32-bit to 64-bit. This affects `struct stat`
  and `struct dirent` structures in Rust libc.
  https://svnweb.freebsd.org/base?view=revision&revision=318736

* Time-related members of `struct kevent` were changed to 64-bit. Again,
  this affects the definition of the same structure in Rust libc.
  https://svnweb.freebsd.org/base?view=revision&revision=320043

Currently, a Rust program compiled on FreeBSD 11.x will work on FreeBSD
12.x thanks to symbol versioning. Unfortunately, a program compiled on
FreeBSD 12.x will break.

Until Rust gains support for target versions, we need to detect the
correct structures to use, based on the build host. This won't make it
possible to build a FreeBSD 11 executable on a FreeBSD 12 host, but it
is good enough when the build and target hosts are the same.

This patch introduces a build script which uses freebsd-version(1) to
detect the version. When it is 12 or more, it defines the
`freebsd12_abi` feature.

Modules in `src/unix/bsd/freebsdlike/freebsd` were reorganized. They are
now split based on this ABI version instead of the target architecture.
For instance, `struct stat` is defined once in `freebsd12.rs` for all
architectures.

`struct kevent` grew a new member, `ext` in FreeBSD 12.x. To avoid too
much pain to users of this structure, both flavors (`freebsd11.rs` and
`freebsd12.rs`) now have the `ext` member. In the case of `freebsd11`,
this is a 0-length array. To help initialize the structure, one can use
the `libc::KEVENT_EXT_ZEROED` constant.

Fixes rust-lang/rust#42681.
