+++ src/unix/bsd/freebsdlike/freebsd/mod.rs
@@ -1,8 +1,6 @@
 pub type fflags_t = u32;
 pub type clock_t = i32;
-pub type ino_t = u32;
 pub type lwpid_t = i32;
-pub type nlink_t = u16;
 pub type blksize_t = u32;
 pub type clockid_t = ::c_int;
 pub type sem_t = _sem;
@@ -41,14 +39,6 @@ s! {
         __unused4: ::c_long,
         __unused5: *mut ::c_void,
         pub aio_sigevent: sigevent
-    }
-
-    pub struct dirent {
-        pub d_fileno: u32,
-        pub d_reclen: u16,
-        pub d_type: u8,
-        pub d_namlen: u8,
-        pub d_name: [::c_char; 256],
     }
 
     pub struct jail {
@@ -593,5 +583,18 @@ cfg_if! {
         pub use self::aarch64::*;
     } else {
         // Unknown target_arch
+    }
+}
+
+cfg_if! {
+    if #[cfg(freebsd12_abi)] {
+        // Starting with FreeBSD 12.0-RELEASE, the kernel uses 64-bit
+        // ino_t. This affects `struct stat` and `struct dirent_t` as
+        // well as a few types above.
+        mod freebsd12;
+        pub use self::freebsd12::*;
+    } else {
+        mod freebsd11;
+        pub use self::freebsd11::*;
     }
 }
